+++ supabase-reset-child-comments.sql
@@ -3,6 +3,11 @@
 
 begin;
 
-delete from public.child_comments;
+do $$
+begin
+  if to_regclass('public.child_comments') is not null then
+    delete from public.child_comments;
+  end if;
+end $$;
 
 commit;
